fix: `Installation Setup Complete` event to fire always after the super user is created  (#24137)
## Description
As we are executing the monos in parallel for `Installation Setup
Complete` event using `Mono.when()`. Now if any of the mono throws
exception, subscriptions of other monos will get cancelled
automatically. To avoid this dependency we are moving the
`sendInstallationSetupAnalytics` mono in a separate thread. Also as a
part of this PR we are handling the possible root cause of exceptions in
NetworkUtils class to handle the timeouts and have `unknown` as a
fallback IP address value in error scenarios.

#### PR fixes following issue(s)
Fixes https://github.com/appsmithorg/appsmith/issues/22388

#### Type of change
- Bug fix (non-breaking change which fixes an issue)

## Testing
#### How Has This Been Tested?
- [x] Manual

## Checklist:
#### Dev activity
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] PR is being merged under a feature flag


#### QA activity:
- [ ] [Speedbreak
features](https://github.com/appsmithorg/TestSmith/wiki/Test-plan-implementation#speedbreaker-features-to-consider-for-every-change)
have been covered
- [ ] Test plan covers all impacted features and [areas of
interest](https://github.com/appsmithorg/TestSmith/wiki/Guidelines-for-test-plans/_edit#areas-of-interest)
- [ ] Test plan has been peer reviewed by project stakeholders and other
QA members
- [ ] Manually tested functionality on DP
- [ ] We had an implementation alignment call with stakeholders post QA
Round 2
- [ ] Cypress test cases have been added and approved by SDET/manual QA
- [ ] Added `Test Plan Approved` label after Cypress tests were reviewed
- [ ] Added `Test Plan Approved` label after JUnit tests were reviewed