chore: Import action and action collection in bulk during file import (#24603)
## Description
When we import an application from file for Git, file import, templates
and other features, we import the list of actions one by one. This leads
to transaction timeout and other performance issues when the application
has a large number of actions and action collections. This PR imports
the actions and action collections in bulk so that number of database
queries are significantly reduced.

#### PR fixes following issue(s)
Fixes #24419