chore: Update chromatic Worfklow (#24976)
On WDS PR's merge, we need to create a build for the release branch too
so that chromatic baselines is updated.
This PR updates the workflow to start a build on merge too.
Also, we can now remove build-storybook workflow as that is not needed
anymore since we will be building chroamtic on release also. We can use
`https://release--5f1e6db0ad879d0022744996.chromatic.com` to see the
storybook for release branch.