feat: Add fallback height to lazy code editor fallback codemirror component (#25047)
This PR prevents the CodeEditor from first collapsing and then expanding
as a result of the LazyCodeEditor performance improvement
implementation. Instead we set a fixed height for the LazyCodeEditor
component based on the CodeEditor component that will replace it. This
way, the transition would be less obvious to users.

Before:


![image](https://github.com/appsmithorg/appsmith/assets/37867493/10cb5ac1-b007-4c06-a722-5a628ca47e6d)

After:

![image](https://github.com/appsmithorg/appsmith/assets/37867493/4d756e04-4667-4ecb-8b73-2397d3980e6b)