feat: debugger error navigation (#25505)
## Description

Property pane navigation on click of a widget error or log from the
debugger. Navigation to a field works if the log has the required meta
data here it would be the property path. With the help of the property
pane config and the widget's properties we generate the payload to set
the required panel, tab, section states. Factory class to clearly
separate out the logic for the different entities.

#### PR fixes following issue(s)
Related to https://github.com/appsmithorg/appsmith/issues/16408
Fixes https://github.com/appsmithorg/appsmith/issues/25465
Fixes https://github.com/appsmithorg/appsmith/issues/25462