chore: Add .editorconfig (#27343)
An indentation syntax error on a workflow file caused all CI pipelines
to break and fail. Adding a `.editorconfig` file with the correct
expected indentation settings can help solve these problems. A couple of
other things this will do for us:

1. Trailing whitespace is automatically removed. This causes a lot of
noise diffs in PRs.
2. Newline at end of files is a Linux file expectation and this can
ensure that. This also causes a lot of noise diffs in PRs.

The editors that majority of the team uses today, support EditorConfig
out of the box. See https://editorconfig.org/#pre-installed.

There's a _lot_ of violations in the repo today, so I'm not adding a CI
check right away. Let's check back in a month or two and assess if we
want to look into a CI check.